+++ html-formatter/javascript/src/cli-main.tsx
@@ -28,13 +28,14 @@ class CucumberHtmlStream extends Transform {
   }
 
   public _flush(callback: TransformCallback): void {
-    readFile(__dirname + '/../main.js', (err: Error, js: Buffer) => {
+    readFile(__dirname + '/../main.js', { encoding: 'utf-8' }, (err: Error, js: string) => {
       if (err) return callback(err)
 
       readFile(
         __dirname +
           '/../../node_modules/@cucumber/react/dist/src/styles/cucumber-react.css',
-        (err: Error, css: Buffer) => {
+          { encoding: 'utf-8' },
+        (err: Error, css: string) => {
           if (err) return callback(err)
 
           this.push(`<!DOCTYPE html>
@@ -43,7 +44,7 @@ class CucumberHtmlStream extends Transform {
     <title>Cucumber</title>
     <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
     <style>
-${css.toString('utf8')}
+${css}
     </style>
   </head>
   <body>
@@ -62,7 +63,7 @@ ${css.toString('utf8')}
       window.CUCUMBER_MESSAGES = ${JSON.stringify(this.envelopes)}
     </script>
     <script>
-${js.toString('utf8')}
+${js}
     </script>
   </body>
 </html>