+++ query/CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [4.0.0] - 2020-03-02
+
+### Changed
+
+* Upgraded gherkin
+
+### Fixed
+
 * Report `Status.UNKNOWN` when status is not known
 * Add `gherkin` as a runtime dependency
 
@@ -52,7 +60,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First JavaScript implementation
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/query/v3.0.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/query/v4.0.0...master
+[4.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v3.0.0...query/v4.0.0
 [3.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v2.0.0...query/v3.0.0
 [2.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v1.1.1...query/v2.0.0
 [1.1.1]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v1.1.0...query/v1.1.1